# src/backend/docs/v1/v1_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "# _В основе лежит мой курсовой проект 😎_",
        "title": "Идентификация на КПП",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/employee-photo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Метод для получения фотографии своего профиля",
                "produces": [
                    "image/jpeg",
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Получение фотографии своего профиля",
                "responses": {
                    "200": {
                        "description": "Фотография успешно получена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/http.StatusBadRequest"
                        }
                    },
                    "401": {
                        "description": "Авторизация неуспешна",
                        "schema": {
                            "$ref": "#/definitions/http.StatusUnauthorized"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка получения фотографии",
                        "schema": {
                            "$ref": "#/definitions/http.StatusInternalServerError"
                        }
                    }
                }
            }
        },
...
        "/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Метод для получения профиля",
                "tags": [
                    "employee"
                ],
                "summary": "Получение профиля",
                "responses": {
                    "200": {
                        "description": "Профиль успешно получен",
                        "schema": {
                            "$ref": "#/definitions/user.getProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/http.StatusBadRequest"
                        }
                    },
                    "401": {
                        "description": "Авторизация неуспешна",
                        "schema": {
                            "$ref": "#/definitions/http.StatusUnauthorized"
                        }
                    },
                    "404": {
                        "description": "Профиль не найден",
                        "schema": {
                            "$ref": "#/definitions/http.StatusNotFound"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка заполнения профиля",
                        "schema": {
                            "$ref": "#/definitions/http.StatusInternalServerError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Метод для заполнения профиля",
                "consumes": [
                    "multipart/form-data",
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Заполнение профиля",
                "parameters": [
                    {
                        "type": "string",
                        "name": "documentFields",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "documentType",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "serialNumber",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Заполнения профиля пользователя",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Профиль успешно создан",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/http.StatusBadRequest"
                        }
                    },
                    "401": {
                        "description": "Авторизация неуспешна",
                        "schema": {
                            "$ref": "#/definitions/http.StatusUnauthorized"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка заполнения профиля",
                        "schema": {
                            "$ref": "#/definitions/http.StatusInternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.StatusBadRequest": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Incorrect request body date of birth"
                }
            }
        },
        "http.StatusInternalServerError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Failed to register employee"
                }
            }
        },
        "http.StatusNotFound": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Failed to get info card document"
                }
            }
        },
        "http.StatusUnauthorized": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Cant login employee"
                }
            }
        },
...
        "model.KeyValueField": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "user.getProfileResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "documentFields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.KeyValueField"
                    }
                },
                "documentType": {
                    "type": "string"
                },
                "isConfirmed": {
                    "type": "boolean"
                },
                "serialNumber": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}